requirements.txt
chromadb>=0.5.16
openai>=1.0.0
langchain-openai>=0.1.0
python-dotenv
tqdm
pydantic>=2.0.0

# Azure Blob Storage
azure-storage-blob>=12.0.0

# Document parsing
pypdf>=3.0.0
pandas>=2.0.0
openpyxl>=3.0.0

# API Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
